Change to size t

+++ Unit02-STL-data-structures/homework2b-vector-set/VectorSet.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef>
+
 template<class T>
 class VectorSet {
 public:
@@ -22,7 +24,7 @@ public:
         // return true if item is removed and false if item wasn't in the set
     }
 
-    int size() const {
+    size_t size() const {
         // implement size here
 
         // return the number of items in the set
